+++ rfc9535/test_suite/results.csv
@@ -1,11 +1,11 @@
 Total; Passed; Failed; Date
-687; 650; 25; 2025-03-16 11:54:46
-687; 650; 25; 2025-03-16 11:55:12
-687; 650; 25; 2025-03-16 11:55:25
-687; 650; 25; 2025-03-16 11:56:18
 687; 650; 25; 2025-03-16 11:56:31
 687; 650; 25; 2025-03-16 11:58:52
 687; 650; 25; 2025-03-16 13:02:53
 687; 651; 24; 2025-03-16 19:37:57
 687; 643; 32; 2025-03-17 22:18:52
 687; 651; 24; 2025-03-17 22:21:55
+687; 651; 24; 2025-03-17 22:31:38
+687; 651; 24; 2025-03-17 22:35:34
+687; 652; 23; 2025-03-17 22:59:27
+687; 652; 23; 2025-03-17 23:00:37
